requirements.txt
discord.py == 1.7.3
dotenv
selenium == 3.14.0
requests == 2.23.0
